+++ spec_fast/lib/r_letters/datasets/segments_spec.rb
@@ -15,7 +15,8 @@ describe RLetters::Datasets::Segments do
 
   context 'one block for the dataset, splitting across' do
     before(:each) do
-      @segments = described_class.new(@dataset).segments
+      @analyzer = described_class.new(@dataset)
+      @segments = @analyzer.segments
     end
 
     it 'creates only one block' do
@@ -28,6 +29,13 @@ describe RLetters::Datasets::Segments do
 
     it 'names the block' do
       expect(@segments[0].name).to eq('Block #1/1')
+    end
+
+    it 'sets dfs correctly' do
+      expect(@analyzer.dfs['it']).to eq(2)
+      expect(@analyzer.dfs['was']).to eq(2)
+      expect(@analyzer.dfs['the']).to eq(2)
+      expect(@analyzer.dfs['best']).to eq(2)
     end
   end
 
@@ -51,40 +59,56 @@ describe RLetters::Datasets::Segments do
 
   context 'one block per document, not splitting across' do
     before(:each) do
-      @segments = described_class.new(@dataset, nil, split_across: false).segments
+      @analyzer = described_class.new(@dataset, nil, split_across: false)
+      @segments = @analyzer.segments
     end
 
-      it 'creates two blocks' do
-        expect(@segments.count).to eq(2)
-      end
+    it 'creates two blocks' do
+      expect(@segments.size).to eq(2)
+    end
 
-      it 'puts all the words in each block' do
-        expect(@segments[0].words.count).to eq(119)
-        expect(@segments[1].words.count).to eq(119)
-      end
+    it 'puts all the words in each block' do
+      expect(@segments[0].words.size).to eq(119)
+      expect(@segments[1].words.size).to eq(119)
+    end
 
-      it 'names the blocks' do
-        expect(@segments[0].name).to eq('Block #1/1 (within document doi:10.1234/5678)')
-        expect(@segments[1].name).to eq('Block #1/1 (within document doi:10.2345/6789)')
-      end
+    it 'names the blocks' do
+      expect(@segments[0].name).to eq('Block #1/1 (within document doi:10.1234/5678)')
+      expect(@segments[1].name).to eq('Block #1/1 (within document doi:10.2345/6789)')
+    end
+
+    it 'sets dfs correctly' do
+      expect(@analyzer.dfs['it']).to eq(2)
+      expect(@analyzer.dfs['was']).to eq(2)
+      expect(@analyzer.dfs['the']).to eq(2)
+      expect(@analyzer.dfs['best']).to eq(2)
+    end
   end
 
   context 'four total blocks, splitting across' do
     before(:each) do
       @doc_segmenter = RLetters::Documents::Segments.new(nil, num_blocks: 4)
-      @segments = described_class.new(@dataset, @doc_segmenter).segments
+      @analyzer = described_class.new(@dataset, @doc_segmenter)
+      @segments = @analyzer.segments
     end
 
     it 'creates four blocks' do
-      expect(@segments.count).to eq(4)
+      expect(@segments.size).to eq(4)
     end
 
     it 'splits the words evenly' do
-      expect(@segments.map(&:words).map(&:count)).to match_array([59, 59, 60, 60])
+      expect(@segments.map(&:words).map(&:size)).to match_array([59, 59, 60, 60])
     end
 
     it 'names the blocks' do
       expect(@segments[0].name).to eq('Block #1/4')
+    end
+
+    it 'sets dfs correctly' do
+      expect(@analyzer.dfs['it']).to eq(2)
+      expect(@analyzer.dfs['was']).to eq(2)
+      expect(@analyzer.dfs['the']).to eq(2)
+      expect(@analyzer.dfs['best']).to eq(2)
     end
   end
 
@@ -92,19 +116,27 @@ describe RLetters::Datasets::Segments do
     before(:each) do
       @doc_segmenter = RLetters::Documents::Segments.new(nil, block_size: 10,
                                                               last_block: :truncate_all)
-      @segments = described_class.new(@dataset, @doc_segmenter).segments
+      @analyzer = described_class.new(@dataset, @doc_segmenter)
+      @segments = @analyzer.segments
     end
 
     it 'creates only one block' do
-      expect(@segments.count).to eq(1)
+      expect(@segments.size).to eq(1)
     end
 
     it 'puts ten words in the block' do
-      expect(@segments[0].words.count).to eq(10)
+      expect(@segments[0].words.size).to eq(10)
     end
 
     it 'names the block' do
       expect(@segments[0].name).to eq('Block #1 of 10 words')
+    end
+
+    it 'sets dfs correctly' do
+      expect(@analyzer.dfs['it']).to eq(2)
+      expect(@analyzer.dfs['was']).to eq(2)
+      expect(@analyzer.dfs['the']).to eq(2)
+      expect(@analyzer.dfs['best']).to eq(2)
     end
   end
 
@@ -112,23 +144,31 @@ describe RLetters::Datasets::Segments do
     before(:each) do
       @doc_segmenter = RLetters::Documents::Segments.new(nil, block_size: 10,
                                                               last_block: :truncate_all)
-      @segments = described_class.new(@dataset,
+      @analyzer = described_class.new(@dataset,
                                       @doc_segmenter,
-                                      split_across: false).segments
+                                      split_across: false)
+      @segments = @analyzer.segments
     end
 
     it 'creates two blocks' do
-      expect(@segments.count).to eq(2)
+      expect(@segments.size).to eq(2)
     end
 
     it 'puts ten words in each block' do
-      expect(@segments[0].words.count).to eq(10)
-      expect(@segments[1].words.count).to eq(10)
+      expect(@segments[0].words.size).to eq(10)
+      expect(@segments[1].words.size).to eq(10)
     end
 
     it 'names the blocks' do
       expect(@segments[0].name).to eq('Block #1 of 10 words (within document doi:10.1234/5678)')
       expect(@segments[1].name).to eq('Block #1 of 10 words (within document doi:10.2345/6789)')
+    end
+
+    it 'sets dfs correctly' do
+      expect(@analyzer.dfs['it']).to eq(2)
+      expect(@analyzer.dfs['was']).to eq(2)
+      expect(@analyzer.dfs['the']).to eq(2)
+      expect(@analyzer.dfs['best']).to eq(2)
     end
   end
 end